requirements.txt
av==14.0.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.3.1
Deprecated==1.2.15
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
fsspec==2024.12.0
future==1.0.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
jaconv==0.4.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.0.2
onnxruntime==1.19.2
packaging==24.2
protobuf==5.29.3
pydub==0.25.1
pykakasi==2.3.0
PyYAML==6.0.2
requests==2.32.3
sympy==1.13.3
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
XlsxWriter==3.2.0
zipp==3.21.0
